+++ Kabinett/Presentation/ViewModel/WriteLetter/ImagePickerViewModel.swift
@@ -64,3 +64,15 @@ final class ImagePickerViewModel: ObservableObject {
         }
     }
 }
+
+// MARK: - ViewModel
+final class CameraViewModel: ObservableObject {
+    @Published var capturedImage: UIImage?
+    
+    // MARK: Method(캡쳐된 이미지 처리)
+    func captureImage(with info: [UIImagePickerController.InfoKey : Any]) {
+        if let image = info[.originalImage] as? UIImage {
+            capturedImage = image
+        }
+    }
+}
